Fixes translator trimming bug, adds external editor preference.
When the "Use external editor" box in the Code pane is checked,
the Tests and Code panes are set to read-only, and the translator code
will be loaded afresh whenever _run(..) is called (i.e., for the
doWeb, doImport buttons). This makes it possible to edit the
translator file in the data directory using the text editor of your
choice, but still use Scaffold to test the translator.