Zotero 5.0 compatibility
There might be some additional bugs, but this seems to mostly work.

We're no longer building Zotero for Firefox, but we're maintaining
Firefox compatibility for the time being (since some things remain
easier to debug in Firefox), so this can still be used with a source
install of Zotero 5.0 [1].

With some minor changes (e.g., not looking for an active browser tab),
we can probably get this working in Standalone.

Closes #34

[1] https://www.zotero.org/support/dev/source_code#running_git_installations