Use an external translators directory
Require a translators directory to be specified at startup, save to that
directory by default, and add an additional button to save to the
current Zotero instance.

There's more that could be done here, such as checking the current
translator for changes when regaining focus or adding a menu option to
reload all translators (in case external changes are made to a
translator other than the current one).

This also updates toolbar buttons to 2x (without conditional sizing, so
low-res screens may be a little blurry) and changes the import buttons
to be a bit more logical, if not great works of art.

Addresses #105 -- this will work as soon as zotero/zotero#1711 lands